feat: extract `Report` into `ReportBuilder`
BREAKING CHANGE: Renamed `Report` to `ReportBuilder` as it's more accurate
